case-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Backend"
authors = ["gsifdgs <ibrahimfaruk1@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
djangorestframework = "^3.15.2"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.7.1"
uvicorn = "^0.30.5"
django-environ = "^0.11.2"
requests = "^2.32.3"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
